made second function return a long int

+++ recursion/5-sqrt_recursion.c
@@ -20,7 +20,7 @@ int _sqrt_recursion(int n)
  * Return: the natural square root of a number
  * If it is a decimal, the number will not be found, so it returns -1
  */
-int _sqrt_part_two(int n, int m)
+long int _sqrt_part_two(int n, int m)
 {
 	if (m * m == n)
 		return (m);
